refactor: middleware -> proxy
