+++ frontend/core/proxies/avoid-scan.ts
