+++ frontend/core/proxies/query-whitelist.ts
