+++ frontend/core/tsconfig.json
@@ -46,8 +46,8 @@
       "~/const/*": ["./constant/*"],
       "~/providers": ["../core/providers/"],
       "~/providers/*": ["../core/providers/*"],
-      "~/middlewares": ["../core/middlewares/"],
-      "~/middlewares/*": ["../core/middlewares/*"],
+      "~/proxies": ["../core/proxies/"],
+      "~/proxies/*": ["../core/proxies/*"],
       "~/tailwind/*": ["./tailwind/*"],
       "~/tailwind": ["./tailwind"]
     }
