+++ frontend/gateway/proxy.ts
@@ -1,18 +1,18 @@
-import { NextResponse } from 'next/server'
 import type { NextRequest } from 'next/server'
+import { NextResponse } from 'next/server'
 
 import { includes } from 'ramda'
 
 import {
+  getDashboardUrl,
+  isDashboardRoute,
+  isDashboardStaticRoute,
+  isLandingStaticRoute,
   SITE,
   STATIC_PATHS,
-  isLandingStaticRoute,
-  isDashboardStaticRoute,
-  isDashboardRoute,
-  getDashboardUrl,
 } from './utils'
 
-export default function middleware(request: NextRequest) {
+export default function proxy(request: NextRequest) {
   const url = request.nextUrl
   const { pathname, host, search } = url
